fix lỗi không xóa được môn trong khóa học

+++ src/main/java/medical/education/service/CourseSubjectServiceImpl.java
@@ -47,6 +47,9 @@ public class CourseSubjectServiceImpl extends
 
   @Override
   public void delete(Long courseId, Long subjectId) {
+    if (!courseSubjectRepository.exist(courseId,subjectId))
+      throw new BaseException(400,"subject not on course");
+
     getRepository().deleteByCourseIdAndSubjectId(courseId,subjectId);
   }
 }
